LPS-97897 Move configuration xml parsing logic to JGroupsClusterChannelFactory
