+++ modules/apps/portal/portal-cluster-multiple/src/main/java/com/liferay/portal/cluster/multiple/internal/jgroups/JGroupsClusterChannelFactory.java
@@ -15,15 +15,19 @@
 package com.liferay.portal.cluster.multiple.internal.jgroups;
 
 import com.liferay.petra.concurrent.ConcurrentReferenceKeyHashMap;
+import com.liferay.petra.io.StreamUtil;
+import com.liferay.petra.io.unsync.UnsyncByteArrayInputStream;
 import com.liferay.petra.memory.FinalizeManager;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.cluster.multiple.configuration.ClusterExecutorConfiguration;
 import com.liferay.portal.cluster.multiple.internal.ClusterChannel;
 import com.liferay.portal.cluster.multiple.internal.ClusterChannelFactory;
 import com.liferay.portal.cluster.multiple.internal.ClusterReceiver;
 import com.liferay.portal.cluster.multiple.internal.io.ClusterClassLoaderPool;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
@@ -31,15 +35,22 @@ import com.liferay.portal.kernel.util.Html;
 import com.liferay.portal.kernel.util.Props;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.SocketUtil;
+import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 
+import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.io.InputStream;
 
 import java.net.InetAddress;
 import java.net.NetworkInterface;
 
 import java.util.Map;
+import java.util.Properties;
 import java.util.concurrent.ConcurrentMap;
+
+import org.jgroups.conf.ConfiguratorFactory;
+import org.jgroups.conf.ProtocolStackConfigurator;
 
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
@@ -66,10 +77,16 @@ public class JGroupsClusterChannelFactory implements ClusterChannelFactory {
 		String channleLogicName, String channelProperties, String clusterName,
 		ClusterReceiver clusterReceiver) {
 
-		return new JGroupsClusterChannel(
-			channleLogicName, channelProperties, clusterName, clusterReceiver,
-			_bindInetAddress, _clusterExecutorConfiguration, _classLoaders,
-			_props, _html);
+		try {
+			return new JGroupsClusterChannel(
+				channleLogicName, _parseProperties(channelProperties),
+				clusterName, clusterReceiver, _bindInetAddress,
+				_clusterExecutorConfiguration, _classLoaders);
+		}
+		catch (Exception e) {
+			throw new SystemException(
+				"Unable to create JGroupsClusterChannel", e);
+		}
 	}
 
 	@Override
@@ -228,6 +245,41 @@ public class JGroupsClusterChannelFactory implements ClusterChannelFactory {
 	@Reference(unbind = "-")
 	protected void setProps(Props props) {
 		_props = props;
+	}
+
+	private ProtocolStackConfigurator _parseProperties(String configXMLFile)
+		throws Exception {
+
+		try (InputStream inputStream = ConfiguratorFactory.getConfigStream(
+				configXMLFile)) {
+
+			if (inputStream == null) {
+				throw new FileNotFoundException(
+					"Config file not found: ".concat(configXMLFile));
+			}
+
+			String configXML = StreamUtil.toString(inputStream);
+
+			Properties properties = _props.getProperties();
+
+			for (Map.Entry<Object, Object> entry : properties.entrySet()) {
+				if (!(entry.getValue() instanceof String)) {
+					continue;
+				}
+
+				configXML = StringUtil.replace(
+					configXML,
+					StringBundler.concat(
+						StringPool.DOLLAR_AND_OPEN_CURLY_BRACE,
+						_html.escapeAttribute((String)entry.getKey()),
+						StringPool.CLOSE_CURLY_BRACE),
+					_html.escapeAttribute((String)entry.getValue()));
+			}
+
+			return ConfiguratorFactory.getStackConfigurator(
+				new UnsyncByteArrayInputStream(
+					configXML.getBytes(StringPool.UTF8)));
+		}
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
